make rails 5.2 use the correct rails version, and fix some tests
Why is this change needed?
--------------------------
Messed up the setup with rails 5.2 where it installed rails 5.1. Also
there is a regression in the sqlite3 gem that was not backported to
rails 4x, so we need to conditionalize the installation of that per
rails version.

This also conditionalizes some sql that is output by the gem for rails
5.2 since it changes slightly, and removes an old conditional around
spaces since we are not supporting 4.1 any more.

How does it address the issue?
------------------------------
Just fix gemfiles.

Any links to any relevant tickets, articles or other resources?
---------------------------------------------------------------
sqlite3 gem issue: https://github.com/rails/rails/issues/35161